requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlite3
passlib[bcrypt]
python-jose